Unpack inner OpenRPA.BusinessRuleException
When a Business Rule Exception is thrown within a sub-workflow, it is returned as a regular System.Exception with an OpenRPA.BusinessRuleException as the inner exception. It contains information, that may be relevant for a technical/application exception (any other exception than a Business Rule Exception), such as in which workflow the error occured, and a stacktrace. This is not really relevant for Business Rule Exceptions. So I propose to unpack the BusinessRuleException.